structure change and testing movement between pages
