+++ code/pom.xml
@@ -1,21 +1,20 @@
-<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-  xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/maven-v4_0_0.xsd">
+<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/maven-v4_0_0.xsd">
 	<modelVersion>4.0.0</modelVersion>
-	
+
 	<!-- project info -->
 	<groupId>Group.Four</groupId>
 	<artifactId>Sample-Code-Coverage-Project</artifactId>
 	<version>1.0-SNAPSHOT</version>
-	
+
 	<properties>
 		<project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
 		<maven.compiler.source>11</maven.compiler.source>
 		<maven.compiler.target>11</maven.compiler.target>
 	</properties>
-	
+
 	<dependencies>
 		<!-- JavaFX -->
-		<dependency> 
+		<dependency>
 			<groupId>org.openjfx</groupId>
 			<artifactId>javafx-controls</artifactId>
 			<version>15.0.1</version>
@@ -38,27 +37,30 @@
 			<version>5.5.1</version>
 			<scope>test</scope>
 		</dependency>
-	
-   <dependency>
-     <groupId>org.json</groupId>
-     <artifactId>json</artifactId>
-     <version>20190722</version>
-   </dependency>
- </dependencies>
+
+		<dependency>
+			<groupId>org.json</groupId>
+			<artifactId>json</artifactId>
+			<version>20190722</version>
+		</dependency>
+	</dependencies>
 	<build>
-		<plugins>        
-			<plugin> <!-- clean lifecycle, see https://maven.apache.org/ref/current/maven-core/lifecycles.html#clean_Lifecycle -->
+		<plugins>
+			<plugin>
+				<!-- clean lifecycle, see https://maven.apache.org/ref/current/maven-core/lifecycles.html#clean_Lifecycle -->
 				<artifactId>maven-clean-plugin</artifactId>
 				<version>3.1.0</version>
 			</plugin>
-			
-			<plugin><!-- compiler services -->
+
+			<plugin>
+				<!-- compiler services -->
 				<groupId>org.apache.maven.plugins</groupId>
 				<artifactId>maven-compiler-plugin</artifactId>
 				<version>3.8.0</version>
 			</plugin>
-			
-			<plugin><!-- testing services -->
+
+			<plugin>
+				<!-- testing services -->
 				<groupId>org.apache.maven.plugins</groupId>
 				<artifactId>maven-surefire-plugin</artifactId>
 				<version>2.22.2</version>
@@ -68,38 +70,43 @@
 				<artifactId>maven-failsafe-plugin</artifactId>
 				<version>2.22.2</version>
 			</plugin>
-			
-			<plugin><!-- javafx support -->
+
+			<plugin>
+				<!-- javafx support -->
 				<groupId>org.openjfx</groupId>
 				<artifactId>javafx-maven-plugin</artifactId>
 				<version>0.0.6</version>
 				<configuration>
-					<mainClass>Main</mainClass>
+					<mainClass>WorkoutManager.Main</mainClass>
 				</configuration>
 			</plugin>
-			
-			<plugin><!-- builds executable jar files -->
+
+			<plugin>
+				<!-- builds executable jar files -->
 				<artifactId>maven-jar-plugin</artifactId>
 				<version>3.0.2</version>
-				<configuration><!-- Add this configuration group to establish the main class for launching the app -->
+				<configuration>
+					<!-- Add this configuration group to establish the main class for launching the app -->
 					<archive>
 						<manifest>
 							<addClasspath>true</addClasspath>
 							<mainClass>
 								<!-- The main class for launching the app -->
-								Main
+								WorkoutManager.Main
 							</mainClass>
 						</manifest>
 					</archive>
 				</configuration>
 			</plugin>
-			
-			<plugin> <!-- default lifecycle, jar packaging: see https://maven.apache.org/ref/current/maven-core/default-bindings.html#Plugin_bindings_for_jar_packaging -->
+
+			<plugin>
+				<!-- default lifecycle, jar packaging: see https://maven.apache.org/ref/current/maven-core/default-bindings.html#Plugin_bindings_for_jar_packaging -->
 				<artifactId>maven-resources-plugin</artifactId>
 				<version>3.0.2</version>
 			</plugin>
-			
-			<plugin><!-- generates code coverage report and requires minimum code coverage when executing test command -->
+
+			<plugin>
+				<!-- generates code coverage report and requires minimum code coverage when executing test command -->
 				<groupId>org.jacoco</groupId>
 				<artifactId>jacoco-maven-plugin</artifactId>
 				<version>0.8.7</version>
@@ -131,7 +138,7 @@
 									</includes-->
 									<excludes>
 										<exclude>view.codebehind</exclude>
-						
+
 									</excludes>
 									<limits>
 										<limit>
@@ -144,11 +151,12 @@
 							</rules>
 						</configuration>
 					</execution>
-					
+
 				</executions>
 			</plugin>
-			
-			<plugin> <!-- checks code styling -->
+
+			<plugin>
+				<!-- checks code styling -->
 				<groupId>org.apache.maven.plugins</groupId>
 				<artifactId>maven-checkstyle-plugin</artifactId>
 				<version>3.1.2</version>
