+++ code/src/main/java/WorkoutManager/Main.java
@@ -1,3 +1,4 @@
+package WorkoutManager;
 
 import java.io.IOException;
 
@@ -28,7 +29,7 @@ public class Main extends Application {
      */
     @Override
     public void start(Stage primaryStage) throws IOException {
-        Parent parent = FXMLLoader.load(getClass().getResource(Main.PREFRENCES_PAGE));
+        Parent parent = FXMLLoader.load(getClass().getResource(Main.LOGIN_PAGE));
         Scene scene = new Scene(parent);
         primaryStage.setTitle(WINDOW_TITLE);
         primaryStage.setScene(scene);
