+++ code/src/main/java/WorkoutManager/resources/WeightLifting.png
