+++ code/src/main/java/WorkoutManager/view/LoginPage.fxml
@@ -8,7 +8,7 @@
 <?import javafx.scene.image.ImageView?>
 <?import javafx.scene.layout.AnchorPane?>
 
-<AnchorPane maxHeight="-Infinity" maxWidth="-Infinity" minHeight="-Infinity" minWidth="-Infinity" prefHeight="342.0" prefWidth="299.0" xmlns="http://javafx.com/javafx/11.0.1" xmlns:fx="http://javafx.com/fxml/1" fx:controller="view.LoginController">
+<AnchorPane maxHeight="-Infinity" maxWidth="-Infinity" minHeight="-Infinity" minWidth="-Infinity" prefHeight="342.0" prefWidth="299.0" xmlns="http://javafx.com/javafx/11.0.1" xmlns:fx="http://javafx.com/fxml/1" fx:controller="WorkoutManager.view.LoginController">
    <children>
       <TextField fx:id="userNameTextfield" layoutX="75.0" layoutY="180.0" promptText="Enter UserName" />
       <TextField fx:id="passwordTextfield" layoutX="75.0" layoutY="226.0" promptText="Enter Password" />
@@ -17,7 +17,7 @@
       <Button fx:id="loginButton" layoutX="127.0" layoutY="261.0" mnemonicParsing="false" onAction="#handleLogin" text="Login" />
       <ImageView disable="true" fitHeight="129.0" fitWidth="98.0" layoutX="106.0" layoutY="25.0" pickOnBounds="true" preserveRatio="true">
          <image>
-            <Image url="file:/c:/Users/wtjra/OneDrive/Desktop/SeProject/UWG-SE2-Spring22-Team4/code/target/classes/resources/WeightLifting.png" />
+            <Image url="file:/c:/Users/wtjra/OneDrive/Desktop/SeProject/UWG-SE2-Spring22-Team4/code/target/classes/WorkoutManager/resources/WeightLifting.png" />
          </image>
          <viewport>
             <Rectangle2D />
