--- code/src/main/java/view/LoginController.java
@@ -1,35 +1,0 @@
-package view;
-
-import javafx.event.ActionEvent;
-import javafx.fxml.FXML;
-import javafx.scene.control.Button;
-import javafx.scene.control.TextField;
-
-public class LoginController {
-
-    @FXML
-    private TextField userNameTextfield;
-
-    @FXML
-    private TextField passwordTextfield;
-
-    @FXML
-    private Button loginButton;
-
-    @FXML
-    void handleLogin(ActionEvent event) {
-
-    }
-
-    @FXML
-    void initialize() {
-        this.bindLogin();
-
-    }
-
-    private void bindLogin() {
-        this.loginButton.disableProperty().bind(
-                this.userNameTextfield.textProperty().isEmpty().or(this.passwordTextfield.textProperty().isEmpty()));
-    }
-
-}
