--- code/src/main/java/view/PreferencePage.fxml
@@ -1,73 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-
-<?import javafx.scene.control.Button?>
-<?import javafx.scene.control.CheckBox?>
-<?import javafx.scene.control.Label?>
-<?import javafx.scene.effect.DropShadow?>
-<?import javafx.scene.layout.AnchorPane?>
-<?import javafx.scene.layout.ColumnConstraints?>
-<?import javafx.scene.layout.GridPane?>
-<?import javafx.scene.layout.RowConstraints?>
-<?import javafx.scene.text.Font?>
-
-
-<AnchorPane maxHeight="-Infinity" maxWidth="-Infinity" minHeight="-Infinity" minWidth="-Infinity" prefHeight="382.0" prefWidth="373.0" xmlns="http://javafx.com/javafx/11.0.1" xmlns:fx="http://javafx.com/fxml/1" fx:controller="view.PreferenceController">
-   <effect>
-      <DropShadow color="#941212" />
-   </effect>
-   <children>
-      <GridPane fx:id="daysGridPane" gridLinesVisible="true" layoutX="43.0" layoutY="67.0" prefHeight="211.0" prefWidth="100.0">
-        <columnConstraints>
-          <ColumnConstraints hgrow="SOMETIMES" maxWidth="159.20001983642578" minWidth="10.0" prefWidth="159.20001983642578" />
-        </columnConstraints>
-        <rowConstraints>
-          <RowConstraints maxHeight="59.59998168945313" minHeight="10.0" prefHeight="29.399981689453128" vgrow="SOMETIMES" />
-          <RowConstraints maxHeight="112.4" minHeight="4.799999999999997" prefHeight="26.799993896484374" vgrow="SOMETIMES" />
-          <RowConstraints maxHeight="155.4" minHeight="0.0" prefHeight="31.80001220703126" vgrow="SOMETIMES" />
-            <RowConstraints maxHeight="155.4" minHeight="10.0" prefHeight="31.79998779296875" vgrow="SOMETIMES" />
-            <RowConstraints maxHeight="155.4" minHeight="10.0" prefHeight="32.60001220703124" vgrow="SOMETIMES" />
-            <RowConstraints maxHeight="155.4" minHeight="10.0" prefHeight="30.20001831054688" vgrow="SOMETIMES" />
-            <RowConstraints maxHeight="155.4" minHeight="10.0" prefHeight="33.399981689453114" vgrow="SOMETIMES" />
-        </rowConstraints>
-         <children>
-            <CheckBox fx:id="mondayCheckBox" mnemonicParsing="false" prefHeight="18.0" prefWidth="78.0" text="Monday" />
-            <CheckBox fx:id="tuesdayCheckbox" mnemonicParsing="false" text="Tuesday" GridPane.rowIndex="1" />
-            <CheckBox fx:id="wednesdayCheckbox" mnemonicParsing="false" text="Wednesday" GridPane.rowIndex="2" />
-            <CheckBox fx:id="thursdayCheckbox" mnemonicParsing="false" text="Thursday" GridPane.rowIndex="3" />
-            <CheckBox fx:id="fridayCheckbox" mnemonicParsing="false" text="Friday" GridPane.rowIndex="4" />
-            <CheckBox fx:id="saturdayCheckbox" mnemonicParsing="false" text="Saturday" GridPane.rowIndex="5" />
-            <CheckBox fx:id="sundayCheckbox" mnemonicParsing="false" text="Sunday" GridPane.rowIndex="6" />
-         </children>
-      </GridPane>
-      <Label alignment="CENTER" layoutX="31.0" layoutY="32.0" prefHeight="27.0" prefWidth="123.0" text="Preferred Days" textFill="RED">
-         <font>
-            <Font size="18.0" />
-         </font>
-      </Label>
-      <GridPane fx:id="muscleGridPane" gridLinesVisible="true" layoutX="202.0" layoutY="67.0" prefHeight="186.0" prefWidth="116.0">
-        <columnConstraints>
-          <ColumnConstraints hgrow="SOMETIMES" minWidth="10.0" prefWidth="100.0" />
-        </columnConstraints>
-        <rowConstraints>
-          <RowConstraints minHeight="10.0" prefHeight="30.0" vgrow="SOMETIMES" />
-          <RowConstraints minHeight="10.0" prefHeight="30.0" vgrow="SOMETIMES" />
-          <RowConstraints minHeight="10.0" prefHeight="30.0" vgrow="SOMETIMES" />
-            <RowConstraints minHeight="10.0" prefHeight="30.0" vgrow="SOMETIMES" />
-            <RowConstraints minHeight="10.0" prefHeight="30.0" vgrow="SOMETIMES" />
-        </rowConstraints>
-         <children>
-            <CheckBox fx:id="armsCheckBox" mnemonicParsing="false" text="Arms" />
-            <CheckBox fx:id="legsCheckbox" mnemonicParsing="false" text="Legs" GridPane.rowIndex="1" />
-            <CheckBox fx:id="backCheckbox" mnemonicParsing="false" text="Back" GridPane.rowIndex="2" />
-            <CheckBox fx:id="coreCheckbox" mnemonicParsing="false" text="Core" GridPane.rowIndex="3" />
-            <CheckBox fx:id="stretchCheckbox" mnemonicParsing="false" text="Stretch" GridPane.rowIndex="4" />
-         </children>
-      </GridPane>
-      <Label alignment="CENTER" layoutX="180.0" layoutY="32.0" prefHeight="27.0" prefWidth="137.0" text="Preferred Muscle" textFill="RED">
-         <font>
-            <Font size="18.0" />
-         </font>
-      </Label>
-      <Button fx:id="applyPrefButton" layoutX="118.0" layoutY="308.0" mnemonicParsing="false" onAction="#handleApplyPrefButton" prefHeight="26.0" prefWidth="138.0" text="Apply Preferences" />
-   </children>
-</AnchorPane>
